virt: Replace generic VpHaltReasons with new CpuIo::fatal_error, add configurable policy (#1911)
The intention behind this change is to provide better crash stacks and
dumps in crash reports. By panicking at the point of error construction,
rather than after the error has bubbled up to a higher layer, we are
better able to capture stack context around the cause of the failure and
to provide a meaningful stack to our atomations (instead of just blaming
the HaltRequest::Panic handler for all these different error sources).
Essentially simulating 'unwrapping' these errors.

However the ability to keep a VM 'alive' for inspection after a failure
is still useful, and we want to preserve it. Therefore this change makes
this new behavior configurable.