vmbus: Make all members of SavedStateData public (#2033)
For a closed-source save/restore change, we need to be able to manually
construct a vmbus saved state from another crate. This change publics
all of the types needed to construct a `SavedStateData`.

Also included is a bugfix for `proxyintegration` to ensure that only
`Accepted` GPADLs are sent to the proxy. Not doing this can cause
duplicate GPADLs to be sent.