minimal_rt: fix intrinsics to correctly return original pointer (#2041)
The implementation for intrinsics in minimal_rt was incorrect. For
example, the inline asm block for `memcpy` correctly describes that
`rdi` will be clobbered by `rep movsb`, but `memcpy`'s API requires that
the return value is the _original_ destination pointer, not the
incremented pointer. Fix this by removing the mut on `dest` and
indicating to the compiler to discard the clobbered value.

It seems that in most cases, the compiler ignored the return value of
memcpy, except when #2031 was being added. Certain compilation options
or function implementations would cause the compiler to use the inlined
returned value of `memcpy` directly from the call to realloc, which
would then cause a later an assertion failure due to the pointer not
being what was expected and overlapping with the new destination.

Fix all the intrinsics in `minimal_rt` to correctly return the
unmodified `dest` or `ptr` as they should.