petri: Move the vmm test timeout to be entirely within Petri's control (#2016)
We're currently experiencing an issue on our SNP test runners where a
hyper-v VM is taking upwards of 5 minutes to create, which causes us to
hit our test timeout. The thing is, we don't really care about how long
hyper-v takes to create the VM, we only care about how long the VM is
running for, since that's when it's actually executing test content.

The problem is nextest has no visibility into what portion of the code
being run is "setup" and what is actual test content. To solve this,
move the vmm test timeout to be handled entirely within Petri, and
started only once the VM is started. This also removes the previous
issue of having to duplicate the timeout duration between nextest and
petri.

We still keep a timeout in nextest as a backup, just in case the test
completely hangs.