vmbus_server: allow proxy driver to specify a device order (#2038)
This change allows vmbusproxy to specify a device order that will be
used to determine the offer order for offers provided by
`proxyintegration`. This means that in combination with the
`use_absolute_channel_order` option, vmbusproxy can control the order in
which devices get assigned channel IDs.

This is not a breaking change; although the
`VMBUS_PROXY_NEXT_ACTION_OUTPUT` structure has increased in size, older
versions of vmbusproxy.sys which don't support device order will simply
not write to the new field, which is therefore left at zero.