Add a preparation step to run before VMM tests to allow using pipette in a CVM (#2051)
This PR adds a new binary inside of vmm_tests named prep_steps (open to
bikeshedding). This binary, when run in the appropriate environment,
will use petri's automation to boot a Windows UEFI VM, mount a second
Windows image as a data disk, and inject the pipette registry entries
into this second Windows image. This is necessary because our normal
method of injecting these registry entries, the IMC device, is
restricted in a CVM.

Note that this prep_steps tool is run locally on the vmm_test machine,
rather than ahead-of-time. While this does mean the work is duplicated,
this is necessary to avoid artifacting a 30GB VHD between steps.

This PR then adds the necessary infrastructure for consuming this newly
prepared image in our vmm_tests. It moves the CVM boot tests from
no_agent to yes_agent, to verify that pipette is working correctly.

I also removed a large number of boot_no_agent tests that already have
normal boot_with_agent coverage, to try and free up a little CI time.