vmgstool: add test helper and integration tests (#2029)
Add test helper to VmgsTool that creates a VMGS file with two keys. This
can happen if update-key successfully writes a new key but is
interrupted before removing the old one. This allows for the recovery
path in update-key (added in
https://github.com/microsoft/openvmm/commit/1fd1b42ae46c61761bb543aab7f861ab92416a6d)
to be tested synthetically.

Also add vmgstool integration tests that create or modify a vmgs file
using vmgstool and then try to boot the vm.