petri: Increase test timeout (#2079)
We're seeing some tests that look like they're still active from the
logs but are timing out. When I moved the vmm test timeout to be solely
in petri's control instead of shared with nextest I lowered the timeout
to 5 minutes from 6, hoping that that change would allows us to get a
little time back. It seems I was overly optimistic. Raise the timeout
back to 6 minutes.