firmware_uefi: EfiDiagnostics adjust trace channel handling (#2054)
This change simply adjusts it so that UEFI logs that _contain_ ERROR or
WARNING get sent to the tracing::error!() or tracing::warn!() channels
respectively. This is different from the current behavior, where logs
that are EXACTLY ERROR or WARNING would only get sent to the respective
trace macros.

It also moves the string conversion logic to the processing function.