vmm_tests: increase memory validation test threshold (#2093)
Description:
This PR modifies the VMM memory validation tests by increasing the upper
threshold of the baseline values that test results are compared against.
This is done to avoid the need to rerun tests that have high variance.
The methodology of these tests is now to flag any significant/alarming
changes in memory usage.

Changes:
- Set threshold to 1MiB for general memory usage and underhill_vm
process usage on 2VP tests
- Set threshold to 3MiB for general memory usage and underhill_vm
process usage on 64VP tests
- Set threshold to 250kB for all other process metrics
- Added comments providing context for threshold values