CI: Don't require ARM VMM tests to pass right now (#2105)
There is a bug somewhere in ARM's UEFI that is causing VMM tests to
randomly fail with spurious VMBus errors. The investigation of this bug
is ongoing. However, until we have a fix, we should unblock CI for
others. Don't require these tests to pass for the moment to allow
merges, but still run them.